logout checking

+++ frontend/src/redux/operations/authOperations.js
@@ -1,5 +1,6 @@
 import authAction from '../actions/authAction';
 import axios from 'axios';
+import { authSelector } from '../selectors';
 
 axios.defaults.baseURL = 'http://financial-assistant-bc22.herokuapp.com';
 
@@ -39,7 +40,12 @@ const userLogin = credentials => dispatch => {
     });
 };
 
-const userLogout = () => dispatch => {
+const userLogout = () => (dispatch, getState) => {
+  const persistedToken = authSelector.isAuthenticated(getState());
+  if (!persistedToken) {
+    return;
+  }
+  token.set(persistedToken);
   axios.delete('/api/v1/auth/sign-out').then(res => {
     token.unSet();
     dispatch(authAction.logout());
